refactor: move bedrock pricing refresh to connector
